sqrt

+++ recursion/5-sqrt_recursion.c
@@ -5,7 +5,7 @@
  * @m: comparison number
  * Return: 1 if not found sqrroot, else sqrroot
  **/
-long int sqrtRecursive(long int n,long int m)
+long int sqrtRecursive(long int n, long int m)
 {
 if (n <= 0)
 {
@@ -21,8 +21,8 @@ return (sqrtRecursive(n - 1, m));
  * _sqrt_recursion - finds the natural square root of a number
  * @n: given number
  * Return: square root of n or -1
- **/
-long int _sqrt_recursion(long int n)
+**/ 
+int _sqrt_recursion(int n)
 {
 	if (n == 1)
 		return (1);
